get locations on join, add `authenticated?` method
Added `getAllPositions` method to chat handler to populate map with
locations of all currently connected members of a channel. The geo
handler now initializes the map, gets the user's location from the
geolocation API, then calls chat.getAllPositions, which sends a
`location_request` message to all users connected to the channel.
Upon receiving this request, the chat handler sends the position of
the user to the channel in a `location` message.

An `authenticated?` method has been added to the application
controller. This method is employed in the channels, locations, and
messages controller to ensure only authenticated users are allowed to
interact with said controllers. A `401.html` file has been added to
deal with unauthorized requests.