need help
I'm attempting to associate channels with users so that they will act
solely as markers for past channel activity rather than creating and
deleting channels globally. In order to do so I've made the changes you
can see in this commit. After a successful migration I found that I'm
getting a nasty error when trying to create a new channel. See the
error report here:

https://github.com/ngoldman/geochat/issues/1