restore extra options on channels#show
Ditching specific route matches in favor of `resources :channels`
(which was committed later on the master branch) seems to have resolved
the issues encountered in this branch. Needing to reset the database is
less than ideal but since the user base is currently hovering around 1,
I think I can forgive myself.

Note: need to deal with a bad/outdated user ID being stored in the
session cookie. For now it returns a nasty error that can only be
resolved by manually hitting `/signout` or clearing the browser cache.
I'll refrain from merging this branch until this issue is resolved.